change thresh to float

+++ bluedb/db_device_thresh.go
@@ -8,14 +8,14 @@ import (
 )
 
 type DeviceThresh struct {
-	Id             string `orm:"size(64);pk"`
-	DevId          string `orm:"size(64)"` // db in device
-	ProjectId      string `orm:"size(64)"`
-	DeviceId       string `orm:"size(128)"`
-	TemperatureMin int    `orm:"default(0)"`
-	HumidityMin    int    `orm:"default(0)"`
-	TemperatureMax int    `orm:"default(0)"`
-	HumidityMax    int    `orm:"default(0)"`
+	Id             string  `orm:"size(64);pk"`
+	DevId          string  `orm:"size(64)"` // db in device
+	ProjectId      string  `orm:"size(64)"`
+	DeviceId       string  `orm:"size(128)"`
+	TemperatureMin float32 `orm:"default(0)"`
+	HumidityMin    float32 `orm:"default(0)"`
+	TemperatureMax float32 `orm:"default(0)"`
+	HumidityMax    float32 `orm:"default(0)"`
 }
 
 func init() {
